pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mathpilot"
version = "0.1.0"
description = "Scientific workflow assistant - AI-powered algorithm implementation from papers"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "MathPilot Contributors"}]
keywords = ["cli", "scientific", "research", "arxiv", "workflow"]

dependencies = [
    "typer>=0.9.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "tenacity>=8.2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
llm = [
    "anthropic>=0.7.0",
    "openai>=0.27.0",
]
executor = [
    "jupyter>=1.0.0",
    "ipykernel>=6.20.0",
]

[project.scripts]
mathpilot = "mathpilot.cli.main:app"

[tool.setuptools]
packages = ["mathpilot"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=mathpilot --cov-report=term-missing"
